+++ repondeur/zam_repondeur/fetch/__init__.py
@@ -7,8 +7,8 @@ from tlfp.tools.parse_texte import parse
 from zam_aspirateur.amendements.models import Amendement
 from zam_aspirateur.exceptions import NotFound
 
-from zam_repondeur.fetch.an import aspire_an
-from zam_repondeur.fetch.senat import aspire_senat
+from zam_repondeur.fetch.an.amendements import aspire_an
+from zam_repondeur.fetch.senat.amendements import aspire_senat
 from zam_repondeur.models import DBSession, Amendement as AmendementModel, Lecture
